Feat: 카드리스트에서 바로 삭제 가능
